+++ docs/fr_FR/changelog.md
@@ -4,6 +4,9 @@
 >
 >Pour rappel s'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte.
 
+# Current BETA
+- Utilisation du 3rdparty datetimepicker
+
 # Stable - 26/03/2021
 - Support des expressions cron
 - Nouvelle commande pour récupérer la prochaine date de ramassage